project/requirements.txt
flask==2.2.2
flask-sqlalchemy==2.5.1
flask-Security==3.0.0
flask-restful==0.3.9
flask-Migrate==3.1.0
email_validator==1.2.1
bcrypt==3.2.2
pytest==7.1.2
coverage==6.4.4
beautifulsoup4==4.11.1
alembic==1.8.1
